+++ intro-tour-client/src/app/components/location-page/location-page.component.html
@@ -1,0 +1,2 @@
+<br><br><br><br>
+<p>long: {{x}} <br>lat: {{y}}</p>
