--- intro-tour-client/src/app/location-page/location-page.component.html
@@ -1,3 +1,0 @@
-<p>
-  {{location.location.x}}
-</p>
